requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
python-multipart>=0.0.6
aiohttp>=3.9.0
aiofiles>=23.2.0
pyyaml>=6.0
pillow>=10.0.0
httpx>=0.25.0
pydantic>=2.5.0
